Convert jsonrpc-runtime-client from binaries into lib
