+++ circuit/src/rpc-test/jsonrpc-runtime-client/Cargo.toml
@@ -12,7 +12,7 @@ targets = ["x86_64-unknown-linux-gnu"]
 [dependencies]
 futures = "0.1.29"
 hyper = "~0.12.35"
-jsonrpc-core-client = { version = "15.1.0", default-features = false, features = ["http"] }
+jsonrpc-core-client = { version = "15.1.0", features = ["http"] }
 log = "0.4.8"
 node-primitives = { version = "2.0.0", git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.8" }
 sp-tracing = { version = "3.0.0", git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.8" }
