+++ circuit/src/rpc-test/jsonrpc-runtime-client/src/main.rs
@@ -15,22 +15,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#![warn(missing_docs)]
-
-//! Example substrate RPC client code.
-//!
-//! This module shows how you can write a Rust RPC client that connects to a running
-//! substrate node and use statically typed RPC wrappers.
-
-use relay_substrate_client::{Client as SubstrateClient, ConnectionParams};
-
-/// Implement Chain with Polkadot-like types for relay-client
-pub mod polkadot_like_chain;
-pub use polkadot_like_chain::PolkadotLike;
-
-/// Useful Substrate network RPC queries
-pub mod useful_queries;
-pub use useful_queries::get_first_header;
+use jsonrpc_runtime_client::{create_rpc_client, get_first_header, ConnectionParams};
 
 fn main() {
     let sub_params = ConnectionParams {
@@ -45,15 +30,4 @@ fn main() {
         let first_header = get_first_header(&client).await.unwrap();
         println!("first header {:?}", first_header);
     });
-}
-
-/// Run single transaction proof relay and stop.
-pub async fn create_rpc_client(
-    sub_params: &ConnectionParams,
-) -> Result<SubstrateClient<PolkadotLike>, String> {
-    let sub_client = SubstrateClient::<PolkadotLike>::try_connect(sub_params.clone())
-        .await
-        .map_err(|e| e.to_string())?;
-
-    Ok(sub_client)
 }
